feat(dashboard): UI/UX optimization pass from adversarial review
Implement the command-center polish plan: trustworthy hero metrics and
prior-period deltas, dual-axis clarity with low-sample cues, continuous
day axes, sticky header, result mode and theme controls, empty/untagged
states, timezone labeling, chart soft-refresh freeze, and medium/nit
craft (export messaging, group-by guards, a11y, table density).

+++ src/contexts/TimezoneContext.tsx
@@ -93,10 +93,11 @@ export function TimezoneProvider({ children }: { children: React.ReactNode }) {
     
     const config = getTimezoneConfig(selectedTimezone)
     
+    // O35: mon-day style (matches chart axis labels) + time
     return date.toLocaleString('en-US', {
       ...config,
-      month: 'numeric',
-      day: 'numeric', 
+      month: 'short',
+      day: 'numeric',
       year: 'numeric',
       hour: '2-digit',
       minute: '2-digit',
@@ -117,7 +118,7 @@ export function TimezoneProvider({ children }: { children: React.ReactNode }) {
     
     return date.toLocaleDateString('en-US', {
       ...config,
-      month: 'numeric',
+      month: 'short',
       day: 'numeric',
       year: 'numeric'
     })
